requirements.txt
django
requests-html
selenium
transformers
torch
beautifulsoup4
selenium

